api/requirements.txt
flask==1.1.2
flask_cors==3.0.10
Werkzeug==0.16.1
requests
git+https://github.com/cividi/frictionless-geo.git@ckan-limit-sort-filter
frictionless_geojson
# pytest